Finally services are nearly empty :+1:
Moved api services. All validations are now in validator util and all upload functions are in upload util
