+++ util/format/format_test.go
@@ -1,4 +1,4 @@
-package util
+package format
 
 import (
 	"path"
@@ -17,16 +17,16 @@ var _ = func() (_ struct{}) {
 	return
 }()
 
-func TestFormatFilesize(t *testing.T) {
-	format := FormatFilesize(0)
+func TestFileSize(t *testing.T) {
+	format := FileSize(0)
 	if format != "0.0 B" {
 		t.Fatalf("Format of 0 bytes gives %s, expected '0.0 B'", format)
 	}
-	format = FormatFilesize(int64(math.Exp2(0)))
+	format = FileSize(int64(math.Exp2(0)))
 	if format != "1.0 B" {
 		t.Fatalf("Format of 1byte gives %s, expected '1.0 B'", format)
 	}
-	format = FormatFilesize(int64(math.Exp2(10)))
+	format = FileSize(int64(math.Exp2(10)))
 	if format != "1.0 KiB" {
 		t.Fatalf("Format of 1024 bytes gives %s, expected '1.0 KiB'", format)
 	}
@@ -53,34 +53,5 @@ func TestInfoHashToMagnet(t *testing.T) {
 	magnet := InfoHashToMagnet("213d354dd354d534d", "Test", "udp://tracker.doko.moe:6969", "udp://tracker.zer0day.to:1337/announce")
 	if magnetExpected != magnet {
 		t.Fatalf("Magnet URL parsed doesn't give the expected result, have this '%s', want this '%s'", magnet, magnetExpected)
-	}
-}
-
-func TestSafe(t *testing.T) {
-	safeString := map[string]string{
-		"'":                  "&#39;",
-		"&":                  "&amp;",
-		"http://exemple.com": "http://exemple.com",
-	}
-	for key, val := range safeString {
-		safe := Safe(key)
-		if string(safe) != val {
-			t.Errorf("Safe doesn't escape the right values, expected result %s, got %s", key, val)
-		}
-	}
-}
-
-func TestSafeText(t *testing.T) {
-	safeString := map[string]string{
-		"'":                                    "&#39;",
-		"&":                                    "&amp;",
-		"http://exemple.com":                   "http://exemple.com",
-		"<em>test</em><script>lol();</script>": "&lt;em&gt;test&lt;/em&gt;&lt;script&gt;lol();&lt;/script&gt;",
-	}
-	for key, val := range safeString {
-		safe := Safe(key)
-		if string(safe) != val {
-			t.Errorf("Safe doesn't escape the right values, expected result %s, got %s", key, val)
-		}
 	}
 }
