+++ util/format/magnet.go
@@ -1,4 +1,4 @@
-package util
+package format
 
 import (
 	"fmt"
