+++ util/sanitize/markdown.go
@@ -1,4 +1,4 @@
-package util
+package sanitize
 
 import (
 	"bytes"
