+++ util/sanitize/safe.go
@@ -1,4 +1,4 @@
-package util
+package sanitize
 
 import (
 	"html"
