+++ util/validator/torrent/forms.go
@@ -3,17 +3,17 @@ package torrentValidator
 // TorrentRequest struct
 // Same json name as the constant!
 type TorrentRequest struct {
-	Name        string `json:"name,omitempty"`
-	Magnet      string `json:"magnet,omitempty"`
-	Category    string `json:"c"`
-	Remake      bool   `json:"remake,omitempty"`
-	Description string `json:"desc,omitempty"`
-	Status      int    `json:"status,omitempty"`
-	Hidden      bool   `json:"hidden,omitempty"`
-	CaptchaID   string `json:"-"`
-	WebsiteLink string `json:"website_link,omitempty"`
-	SubCategory int    `json:"sub_category,omitempty"`
-	Language    string `json:"language,omitempty"`
+	Name        string   `validate:"required" json:"name,omitempty"`
+	Magnet      string   `json:"magnet,omitempty"`
+	Category    string   `validate:"required" json:"c"`
+	Remake      bool     `json:"remake,omitempty"`
+	Description string   `json:"desc,omitempty"`
+	Status      int      `json:"status,omitempty"`
+	Hidden      bool     `json:"hidden,omitempty"`
+	CaptchaID   string   `json:"-"`
+	WebsiteLink string   `validate:"uri" json:"website_link,omitempty"`
+	SubCategory int      `json:"sub_category,omitempty"`
+	Languages   []string `json:"languages,omitempty"`
 
 	Infohash      string         `json:"hash,omitempty"`
 	CategoryID    int            `json:"-"`
